+++ WatsonMesh/MeshServer.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Net;
+using System.Net.Sockets;
 using System.Text;
 using System.Threading.Tasks;
 using WatsonTcp;
@@ -10,42 +12,61 @@ using WatsonTcp;
 namespace WatsonMesh
 { 
     internal class MeshServer : IDisposable
-    {
-        #region Public-Members
-          
+    { 
         internal event EventHandler<ClientConnectionEventArgs> ClientConnected; 
         internal event EventHandler<ClientConnectionEventArgs> ClientDisconnected; 
         internal event EventHandler<StreamReceivedFromClientEventArgs> MessageReceived; 
         internal Action<string> Logger = null;
-
-        #endregion
-
-        #region Private-Members
-
+         
         private bool _Disposed = false;
         private MeshSettings _Settings;
-        private MeshPeer _Self;
+        private string _Ip = null;
+        private int _Port = -1;
+        private string _IpPort = null;
+        private bool _Ssl = false;
+        private string _PfxCertificateFile = null;
+        private string _PfxCertificatePassword = null;
         private WatsonTcpServer _TcpServer;
-
-        #endregion
-
-        #region Constructors-and-Factories
          
-        internal MeshServer(MeshSettings settings, MeshPeer self)
+        internal MeshServer(MeshSettings settings, string ip, int port, bool ssl, string pfxCertFile, string pfxCertPass)
         {
             if (settings == null) throw new ArgumentNullException(nameof(settings));
-            if (self == null) throw new ArgumentNullException(nameof(self));
+            if (String.IsNullOrEmpty(ip)) throw new ArgumentNullException(nameof(ip));
+            if (port < 0) throw new ArgumentException("Port must be zero or greater.");
 
             _Settings = settings;
-            _Self = self;
+            _Ip = ip;
+            _Port = port;
+            _IpPort = ip + ":" + port;
+            _Ssl = ssl;
+            _PfxCertificateFile = pfxCertFile;
+            _PfxCertificatePassword = pfxCertPass;
 
-            Logger?.Invoke("[MeshServer] Initialized on IP:port " + _Self.IpPort);
+            List<string> localIpAddresses = GetLocalIpAddresses();
+            if (_Ip.Equals("127.0.0.1"))
+            {
+                Logger?.Invoke("[MeshServer] Loopback IP address detected; only connections from local machine will be accepted");
+            }
+            else
+            {
+                if (!localIpAddresses.Contains(_Ip))
+                {
+                    Logger?.Invoke("[MeshServer] Specified IP address '" + _Ip + "' not found in local IP address list:");
+                    foreach (string curr in localIpAddresses) Logger?.Invoke("  " + curr);
+                    throw new ArgumentException("IP address must either be 127.0.0.1 or the IP address of a local network interface.");
+                }
+            }
+
+            if (_Ssl)
+            {
+                Logger?.Invoke("[MeshServer] Initialized TCP server with SSL on IP:port " + _IpPort);
+            }
+            else
+            {
+                Logger?.Invoke("[MeshServer] Initialized TCP server on IP:port " + _IpPort);
+            }
         }
-
-        #endregion
-
-        #region Public-Methods
-
+         
         /// <summary>
         /// Tear down the client and dispose of background workers.
         /// </summary>
@@ -59,27 +80,23 @@ namespace WatsonMesh
 
         internal void Start()
         {
-            string ip = null;
-            int port = -1;
-            Common.ParseIpPort(_Self.IpPort, out ip, out port);
-
-            if (_Self.Ssl)
+            if (_Ssl)
             {
                 _TcpServer = new WatsonTcpServer(
-                    ip, 
-                    port,
-                    _Self.PfxCertificateFile,
-                    _Self.PfxCertificatePassword);
+                    _Ip, 
+                    _Port,
+                    _PfxCertificateFile,
+                    _PfxCertificatePassword);
 
-                Logger?.Invoke("[MeshServer] Initialized TCP server with SSL on IP:port " + _Self.IpPort);
+                Logger?.Invoke("[MeshServer] Starting TCP server with SSL on IP:port " + _IpPort);
             }
             else
             {
                 _TcpServer = new WatsonTcpServer(
-                    ip,
-                    port);
+                    _Ip,
+                    _Port);
 
-                Logger?.Invoke("[MeshServer] Initialized TCP server on IP:port " + _Self.IpPort);
+                Logger?.Invoke("[MeshServer] Starting TCP server on IP:port " + _IpPort);
             }
 
             _TcpServer.AcceptInvalidCertificates = _Settings.AcceptInvalidCertificates; 
@@ -102,11 +119,7 @@ namespace WatsonMesh
             Logger?.Invoke("[MeshServer] Disconnecting client " + ipPort); 
             _TcpServer.DisconnectClient(ipPort);
         }
-
-        #endregion
-
-        #region Private-Methods
-
+         
         protected virtual void Dispose(bool disposing)
         {
             if (_Disposed)
@@ -140,6 +153,18 @@ namespace WatsonMesh
             MessageReceived?.Invoke(this, args);
         }
 
-        #endregion
+        private List<string> GetLocalIpAddresses()
+        {
+            List<string> ret = new List<string>();
+            var host = Dns.GetHostEntry(Dns.GetHostName());
+            foreach (var ip in host.AddressList)
+            {
+                if (ip.AddressFamily == AddressFamily.InterNetwork)
+                {
+                    ret.Add(ip.ToString());
+                }
+            }
+            return ret;
+        }
     }
 }